+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.2.0] - 2026-02-02
+
+### Added
+
+- Code actions for taking ownership of files
+  - "Take ownership as individual" - uses configured `individual` setting
+  - "Take ownership as team" - uses configured `team` setting
+  - "Take ownership as custom" - for ad-hoc owners
+  - "Add to existing entry" variants for files with existing ownership
+- New settings: `individual` and `team` for configuring default owners
+
 ## [0.1.1] - 2026-02-02
 
 Initial release.